Create Digital Clock
In this JavaScript Mini project we will make a Digital clock that will display the current time in Hours, Minutes and second. This JavaScript clock display the time as per Local time zone. We will use HTML and CSS to design this Clock, and for Updating the current time we will use JavaScript Date Object. You can use this Digital clock tutorial to make your Mini JavaScript Project for college or work portfolio.